requirements.txt
openslide-python==1.1.2
scikit-image~=0.19.2
scikit-learn~=1.2.1
numpy~=1.23.5
scipy==1.10.0
matplotlib~=3.7.1
dill==0.3.3
pytest~=7.1.3
importlib-resources
typing-extensions
lazy-property
requests~=2.28.1
Pillow~=9.4.0
setuptools~=65.6.3
shapely~=2.0.1
dask~=2024.1.1
